+++ torch/_C/__init__.pyi.in
@@ -181,6 +181,7 @@ def _jit_pass_metal_optimize_for_mobile(module: 'torch.jit.ScriptModule',
 def _jit_pass_inline(Graph) -> None: ...
 def _jit_pass_constant_propagation(Graph) -> None: ...
 def _jit_get_schemas_for_operator(name :str) -> List[FunctionSchema]: ...
+def _jit_check_alias_annotation(g: Graph, args: Tuple[Any, ...], unqualified_op_name: str): ...
 def _jit_can_fuse_on_cpu() -> _bool: ...
 def _jit_can_fuse_on_gpu() -> _bool: ...
 def _debug_get_fusion_group_inlining() -> _bool: ...
@@ -219,6 +220,7 @@ ResolutionCallback = Callable[[str], Callable[..., Any]]
 def _create_function_from_graph(qualname: str, graph: Graph) -> Graph: ...
 def _debug_set_autodiff_subgraph_inlining(disabled: _bool) -> None: ...
 def _ivalue_tags_match(lhs: ScriptModule, rhs: ScriptModule) -> _bool: ...
+def _jit_assert_is_instance(obj: Any, type: JitType): ...
 def _jit_clear_class_registry() -> None: ...
 def _jit_set_emit_hooks(ModuleHook: Optional[Callable], FunctionHook: Optional[Callable]) -> None: ...
 def _jit_get_emit_hooks() -> Tuple[Callable, Callable]: ...
